Split provider and decorator
