+++ src/components/SmartTextLink.stories.tsx
@@ -1,22 +1,23 @@
 import '../storybook/register';
 
-import { text } from 'sku/@storybook/addon-knobs';
 import { Text } from 'braid-design-system';
 import React from 'react';
+import { text } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
-import { BraidStorybookProvider, withProviders } from '../storybook/provider';
+import { withDecorator } from '../storybook/decorator';
+import { StorybookProvider } from '../storybook/provider';
 
 import { SmartTextLink } from './SmartTextLink';
 
 storiesOf('SmartTextLink', module)
   .add('Custom', () => (
-    <BraidStorybookProvider>
+    <StorybookProvider>
       <Text>
         <SmartTextLink href={text('href', 'https://developer.seek.com/schema')}>
           {text('children', 'SEEK Schema')}
         </SmartTextLink>
       </Text>
-    </BraidStorybookProvider>
+    </StorybookProvider>
   ))
-  .addDecorator(withProviders);
+  .addDecorator(withDecorator);
