+++ src/components/WrapperRenderer.stories.tsx
@@ -4,14 +4,15 @@ import { Text } from 'braid-design-system';
 import React from 'react';
 import { storiesOf } from 'sku/@storybook/react';
 
+import { withDecorator } from '../storybook/decorator';
 import Wrapper from '../storybook/markdown/wrapper.mdx';
-import { BraidStorybookProvider, withProviders } from '../storybook/provider';
+import { StorybookProvider } from '../storybook/provider';
 
 import { WrapperRenderer } from './WrapperRenderer';
 
 storiesOf('WrapperRenderer', module)
   .add('Example', () => (
-    <BraidStorybookProvider>
+    <StorybookProvider>
       <WrapperRenderer document={Wrapper}>
         {({ children }) => (
           <Text>
@@ -19,6 +20,6 @@ storiesOf('WrapperRenderer', module)
           </Text>
         )}
       </WrapperRenderer>
-    </BraidStorybookProvider>
+    </StorybookProvider>
   ))
-  .addDecorator(withProviders);
+  .addDecorator(withDecorator);
