+++ src/storybook/provider.tsx
@@ -1,22 +1,15 @@
 import { BraidLoadableProvider, Card, ContentBlock } from 'braid-design-system';
 import React, { ReactNode } from 'react';
-import { Helmet } from 'react-helmet';
-import { BrowserRouter } from 'react-router-dom';
-import { select, text } from 'sku/@storybook/addon-knobs';
-import { addDecorator } from 'sku/@storybook/react';
+import { select } from 'sku/@storybook/addon-knobs';
 
-import { MdxProvider, ScoobieLink } from '..';
-import { robotoHref, robotoMonoHref } from '../../typography';
-import { DEFAULT_SIZE, SIZES } from '../private/size';
+import { ScoobieLink } from '..';
 
-interface BraidStorybookProviderProps {
+interface Props {
   children: ReactNode;
 }
 
 // TODO: this isn't working when loaded as a Storybook decorator.
-export const BraidStorybookProvider = ({
-  children,
-}: BraidStorybookProviderProps) => (
+export const StorybookProvider = ({ children }: Props) => (
   <BraidLoadableProvider
     linkComponent={ScoobieLink}
     themeName={select(
@@ -39,27 +32,4 @@ export const BraidStorybookProvider = ({
       <Card>{children}</Card>
     </ContentBlock>
   </BraidLoadableProvider>
-);
-
-type DecoratorFunction = Parameters<typeof addDecorator>[0];
-
-export const withProviders: DecoratorFunction = (story) => (
-  <MdxProvider
-    graphqlPlayground={
-      text(
-        'MdxProvider.graphqlPlayground',
-        'https://graphql.seek.com/graphql',
-      ) || undefined
-    }
-    size={select('MdxProvider.size', SIZES, DEFAULT_SIZE)}
-  >
-    <BrowserRouter>
-      <Helmet>
-        <link href={robotoHref} rel="stylesheet" />
-        <link href={robotoMonoHref} rel="stylesheet" />
-      </Helmet>
-
-      {story()}
-    </BrowserRouter>
-  </MdxProvider>
 );
